🐛 Fix int type
The integer type in cartesian.cpp was uint_8, which:
  -  Needs to be included by cstdint
  -  Adds no additional performance from int
Consequently it was replaced by the base int type.

+++ src/absi/cartesian.cpp
@@ -65,7 +65,7 @@ static real_interval minMaxOfProduct(real_interval x, real_interval y)
     ampl::real x1 = std::get<0>(x), x2 = std::get<1>(x), y1 = std::get<0>(y), y2 = std::get<1>(y);
     ampl::real values[4] = {x1 * y1, x1 * y2, x2 * y1, x2 * y2};
     ampl::real rmin = values[0], rmax = values[0];
-    for (uint8_t i = 1; i < 4; i++)
+    for (int i = 1; i < 4; i++)
     {
         if (values[i] < rmin)
         {
